fix(probe): keep appliance runtime metadata bounded
